+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_help.py
@@ -86,6 +86,9 @@ helps['vm create'] = """
         - name: Create a CentOS VM with Managed Service Identity. The VM will have a 'Contributor' role with access to a storage account.
           text: >
              az vm create -n MyVm -g MyResourceGroup --image centos --assign-identity --scope /subscriptions/99999999-1bf0-4dda-aec3-cb9272f09590/MyResourceGroup/myRG/providers/Microsoft.Storage/storageAccounts/storage1
+        - name: Create a VM in an availability zone in the current resource group's region
+          text: >
+             az vm create -n MyVm -g MyResourceGroup --image Centos --zone 1
 """
 
 helps['vmss create'] = """
@@ -130,6 +133,9 @@ helps['vmss create'] = """
         - name: Create a VM scaleset with Managed Service Identity. The VM will have a 'Contributor' Role with access to a storage account.
           text: >
              az vm create -n MyVm -g MyResourceGroup --image centos --assign-identity --scope /subscriptions/99999999-1bf0-4dda-aec3-cb9272f09590/MyResourceGroup/myRG/providers/Microsoft.Storage/storageAccounts/storage1
+        - name: Create a single zone VM scaleset in the current resource group's region
+          text: >
+             az vmss create -n MyVmss -g MyResourceGroup --image Centos --zones 1
 """
 
 helps['vm availability-set create'] = """
@@ -1047,6 +1053,9 @@ helps['disk create'] = """
         - name: Create a managed disk by copying an existing disk or snapshot.
           text: >
             az disk create -g MyResourceGroup -n MyDisk2 --source MyDisk
+        - name: Create a disk in an availability zone in the region of "East US 2"
+          text: >
+            az disk create -n MyDisk -g MyResourceGroup --size-gb 10 --location eastus2 --zone 1
 """
 
 helps['disk list'] = """