+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -3,8 +3,9 @@
 Release History
 ===============
 
-unreleased
+2.0.15 (2017-09-22)
 +++++++++++++++++++
+* `vm/vmss/disk create`: support availability zone
 * `vmss create`: Fixed issue where supplying `--app-gateway ID` would fail.
 * `vm create`: Added `--asgs` support.
 * `vm run-command`: support to run commands on remote VMs